move Run test to runner_test.go
